feat: migrate all functions under function handler (wip)
