+++ misc_functions.go
@@ -1,6 +1,6 @@
 package sprout
 
 // Hello returns a greeting string.
-func (p *FunctionHandler) Hello() string {
-	return "Hello, World!"
+func (fh *FunctionHandler) Hello() string {
+	return "Hello!"
 }
